+++ spec/adhearsion/translator/asterisk/component/asterisk/agi_command_spec.rb
@@ -100,6 +100,42 @@ module Adhearsion
                     expect(original_command.response(0.5)).to eq(Adhearsion::ProtocolError.new.setup(:item_not_found, "Could not find a call with ID #{mock_call.id}", mock_call.id))
                   end
                 end
+
+                context "which is 'Channel does not exist: SIP/nosuchchannel'" do
+                  let(:message) { 'Channel does not exist: SIP/nosuchchannel' }
+
+                  it "should return an :item_not_found event for the call" do
+                    subject.execute
+                    expect(original_command.response(0.5)).to eq(Adhearsion::ProtocolError.new.setup(:item_not_found, "Could not find a call with ID #{mock_call.id}", mock_call.id))
+                  end
+                end
+
+                context "which is 'ExtraChannel does not exist: SIP/nosuchchannel'" do
+                  let(:message) { 'ExtraChannel does not exist: SIP/nosuchchannel' }
+
+                  it "should return an :item_not_found event for the call" do
+                    subject.execute
+                    expect(original_command.response(0.5)).to eq(Adhearsion::ProtocolError.new.setup(:item_not_found, "Could not find a call with ID #{mock_call.id}", mock_call.id))
+                  end
+                end
+
+                context "which is 'Redirect failed, channel not up.'" do
+                  let(:message) { 'Redirect failed, channel not up.' }
+
+                  it "should return an :item_not_found event for the call" do
+                    subject.execute
+                    expect(original_command.response(0.5)).to eq(Adhearsion::ProtocolError.new.setup(:item_not_found, "Could not find a call with ID #{mock_call.id}", mock_call.id))
+                  end
+                end
+
+                context "which is 'Redirect failed, channel not up.'" do
+                  let(:message) { 'Redirect failed, extra channel not up.' }
+
+                  it "should return an :item_not_found event for the call" do
+                    subject.execute
+                    expect(original_command.response(0.5)).to eq(Adhearsion::ProtocolError.new.setup(:item_not_found, "Could not find a call with ID #{mock_call.id}", mock_call.id))
+                  end
+                end
               end
             end